Crate egui_plot

Simple plotting library for egui.

Feature flags

Structs

Arrows

A set of arrows.

AxisBools

Two bools, one for each axis (X and Y).

AxisHints

Axis configuration.

Bar

One bar in a BarChart. Potentially floating, allowing stacked bar charts. Width can be changed to allow variable-width histograms.

BarChart

A bar chart.

BoxElem

A box in a BoxPlot diagram. This is a low level graphical element; it will not compute quartiles and whiskers, letting one use their preferred formula. Use Points to draw the outliers.

BoxPlot

A diagram containing a series of BoxElem elements.

BoxSpread

Contains the values of a single box in a box plot.

CoordinatesFormatter

Specifies the coordinates formatting when passed to Plot::coordinates_formatter.

GridInput

Input for “grid spacer” functions.

GridMark

One mark (horizontal or vertical line) in the background grid of a plot.

HLine

A horizontal line in a plot, filling the full width

Legend

The configuration for a plot legend.

Line

A series of values forming a path.

Plot

A 2D plot, e.g. a graph of a function.

PlotBounds

2D bounding box of f64 precision. The range of data values we show.

PlotImage

An image in the plot.

PlotPoint

A point coordinate in the plot.

PlotResponse

What Plot::show returns.

PlotTransform

Contains the screen rectangle and the plot bounds and provides methods to transform between them.

PlotUi

Provides methods to interact with a plot while building it. It is the single argument of the closure provided to Plot::show. See Plot for an example of how to use it.

Points

A set of points.

Polygon

A convex polygon.

Text

Text inside the plot.

VLine

A vertical line in a plot, filling the full width

Enums

Axis

X or Y axis.

Corner

Where to place the plot legend.

HPlacement

Placement of the vertical Y-Axis.

LineStyle

Solid, dotted, dashed, etc.

MarkerShape

Circle, Diamond, Square, Cross, …

Orientation

Determines whether a plot element is vertically or horizontally oriented.

Placement

Placement of an axis.

PlotPoints

Represents many PlotPoints.

VPlacement

Placement of the horizontal X-Axis.

Functions

color_from_strength

Determine a color from a 0-1 strength value.

format_number

Helper for formatting a number so that we always show at least a few decimals, unless it is an integer, in which case we never show any decimals.

log_grid_spacer

Recursively splits the grid into base subdivisions (e.g. 100, 10, 1).

uniform_grid_spacer

Splits the grid into uniform-sized spacings (e.g. 100, 25, 1).