egui_plot

Enum PlotPoints

pub enum PlotPoints<'a> {
    Owned(Vec<PlotPoint>),
    Generator(ExplicitGenerator<'a>),
    Borrowed(&'a [PlotPoint]),
}

Represents many PlotPoints.

These can be an owned Vec or generated on-the-fly by a function or borrowed from a slice.

Variants

Owned(Vec<PlotPoint>)

Generator(ExplicitGenerator<'a>)

Borrowed(&'a [PlotPoint])

Implementations

impl<'a> PlotPoints<'a>

pub fn new(points: Vec<[f64; 2]>) -> Self

pub fn points(&self) -> &[PlotPoint]

pub fn from_explicit_callback( function: impl Fn(f64) -> f64 + 'a, x_range: impl RangeBounds<f64>, points: usize, ) -> Self

Draw a line based on a function y=f(x), a range (which can be infinite) for x and the number of points.

pub fn from_parametric_callback( function: impl Fn(f64) -> (f64, f64), t_range: impl RangeBounds<f64>, points: usize, ) -> Self

Draw a line based on a function (x,y)=f(t), a range for t and the number of points. The range may be specified as start..end or as start..=end.

pub fn from_ys_f32(ys: &[f32]) -> Self

From a series of y-values. The x-values will be the indices of these values

pub fn from_ys_f64(ys: &[f64]) -> Self

From a series of y-values. The x-values will be the indices of these values