egui_plot

Struct BoxPlot

pub struct BoxPlot { /* private fields */ }

A diagram containing a series of BoxElem elements.

Implementations

impl BoxPlot

pub fn new(boxes: Vec<BoxElem>) -> Self

Create a plot containing multiple boxes. It defaults to vertically oriented elements.

pub fn color(self, color: impl Into<Color32>) -> Self

Set the default color. It is set on all elements that do not already have a specific color. This is the color that shows up in the legend. It can be overridden at the element level (see BoxElem). Default is Color32::TRANSPARENT which means a color will be auto-assigned.

pub fn name(self, name: impl ToString) -> Self

Name of this box plot diagram.

This name will show up in the plot legend, if legends are turned on. Multiple series may share the same name, in which case they will also share an entry in the legend.

pub fn vertical(self) -> Self

Set all elements to be in a vertical orientation. Argument axis will be X and values will be on the Y axis.

pub fn horizontal(self) -> Self

Set all elements to be in a horizontal orientation. Argument axis will be Y and values will be on the X axis.

pub fn highlight(self, highlight: bool) -> Self

Highlight all plot elements.

pub fn allow_hover(self, hovering: bool) -> Self

Allowed hovering this item in the plot. Default: true.

pub fn element_formatter( self, formatter: Box<dyn Fn(&BoxElem, &Self) -> String>, ) -> Self

Add a custom way to format an element. Can be used to display a set number of decimals or custom labels.

pub fn id(self, id: Id) -> Self

Set the box plot’s id which is used to identify it in the plot’s response.