egui_plot

Struct Plot

pub struct Plot<'a> { /* private fields */ }

A 2D plot, e.g. a graph of a function.

Plot supports multiple lines and points.

use egui_plot::{Line, Plot, PlotPoints};

let sin: PlotPoints = (0..1000).map(|i| {
    let x = i as f64 * 0.01;
    [x, x.sin()]
}).collect();
let line = Line::new(sin);
Plot::new("my_plot").view_aspect(2.0).show(ui, |plot_ui| plot_ui.line(line));

Implementations

impl<'a> Plot<'a>

pub fn new(id_source: impl Hash) -> Self

Give a unique id for each plot within the same Ui.

pub fn id(self, id: Id) -> Self

Set an explicit (global) id for the plot.

This will override the id set by Self::new.

This is the same Id that can be used for PlotMemory::load.

pub fn data_aspect(self, data_aspect: f32) -> Self

width / height ratio of the data. For instance, it can be useful to set this to 1.0 for when the two axes show the same unit. By default the plot window’s aspect ratio is used.

pub fn view_aspect(self, view_aspect: f32) -> Self

width / height ratio of the plot region. By default no fixed aspect ratio is set (and width/height will fill the ui it is in).

pub fn width(self, width: f32) -> Self

Width of plot. By default a plot will fill the ui it is in. If you set Self::view_aspect, the width can be calculated from the height.

pub fn height(self, height: f32) -> Self

Height of plot. By default a plot will fill the ui it is in. If you set Self::view_aspect, the height can be calculated from the width.

pub fn min_size(self, min_size: Vec2) -> Self

Minimum size of the plot view.

pub fn show_x(self, show_x: bool) -> Self

Show the x-value (e.g. when hovering). Default: true.

pub fn show_y(self, show_y: bool) -> Self

Show the y-value (e.g. when hovering). Default: true.

pub fn center_x_axis(self, on: bool) -> Self

Always keep the X-axis centered. Default: false.

pub fn center_y_axis(self, on: bool) -> Self

Always keep the Y-axis centered. Default: false.

pub fn allow_zoom<T>(self, on: T) -> Self
where T: Into<Vec2b>,

Whether to allow zooming in the plot. Default: true.

Note: Allowing zoom in one axis but not the other may lead to unexpected results if used in combination with data_aspect.

pub fn allow_scroll<T>(self, on: T) -> Self
where T: Into<Vec2b>,

Whether to allow scrolling in the plot. Default: true.

pub fn allow_double_click_reset(self, on: bool) -> Self

Whether to allow double clicking to reset the view. Default: true.

pub fn set_margin_fraction(self, margin_fraction: Vec2) -> Self

Set the side margin as a fraction of the plot size. Only used for auto bounds.

For instance, a value of 0.1 will add 10% space on both sides.

pub fn allow_boxed_zoom(self, on: bool) -> Self

Whether to allow zooming in the plot by dragging out a box with the secondary mouse button.

Default: true.

pub fn boxed_zoom_pointer_button( self, boxed_zoom_pointer_button: PointerButton, ) -> Self

Config the button pointer to use for boxed zooming. Default: Secondary

pub fn allow_drag<T>(self, on: T) -> Self
where T: Into<Vec2b>,

Whether to allow dragging in the plot to move the bounds. Default: true.

pub fn label_formatter( self, label_formatter: impl Fn(&str, &PlotPoint) -> String + 'a, ) -> Self

Provide a function to customize the on-hover label for the x and y axis

use egui_plot::{Line, Plot, PlotPoints};
let sin: PlotPoints = (0..1000).map(|i| {
    let x = i as f64 * 0.01;
    [x, x.sin()]
}).collect();
let line = Line::new(sin);
Plot::new("my_plot").view_aspect(2.0)
.label_formatter(|name, value| {
    if !name.is_empty() {
        format!("{}: {:.*}%", name, 1, value.y)
    } else {
        "".to_owned()
    }
})
.show(ui, |plot_ui| plot_ui.line(line));

pub fn coordinates_formatter( self, position: Corner, formatter: CoordinatesFormatter<'a>, ) -> Self

Show the pointer coordinates in the plot.

pub fn x_grid_spacer( self, spacer: impl Fn(GridInput) -> Vec<GridMark> + 'a, ) -> Self

Configure how the grid in the background is spaced apart along the X axis.

Default is a log-10 grid, i.e. every plot unit is divided into 10 other units.

The function has this signature:

fn step_sizes(input: GridInput) -> Vec<GridMark>;

This function should return all marks along the visible range of the X axis. step_size also determines how thick/faint each line is drawn. For example, if x = 80..=230 is visible and you want big marks at steps of 100 and small ones at 25, you can return:

vec![
   // 100s
   GridMark { value: 100.0, step_size: 100.0 },
   GridMark { value: 200.0, step_size: 100.0 },

   // 25s
   GridMark { value: 125.0, step_size: 25.0 },
   GridMark { value: 150.0, step_size: 25.0 },
   GridMark { value: 175.0, step_size: 25.0 },
   GridMark { value: 225.0, step_size: 25.0 },
];

There are helpers for common cases, see log_grid_spacer and uniform_grid_spacer.

pub fn y_grid_spacer( self, spacer: impl Fn(GridInput) -> Vec<GridMark> + 'a, ) -> Self

Default is a log-10 grid, i.e. every plot unit is divided into 10 other units.

See Self::x_grid_spacer for explanation.

pub fn grid_spacing(self, grid_spacing: impl Into<Rangef>) -> Self

Set when the grid starts showing.

When grid lines are closer than the given minimum, they will be hidden. When they get further apart they will fade in, until the reaches the given maximum, at which point they are fully opaque.

pub fn clamp_grid(self, clamp_grid: bool) -> Self

Clamp the grid to only be visible at the range of data where we have values.

Default: false.

pub fn sense(self, sense: Sense) -> Self

Set the sense for the plot rect.

Default: Sense::click_and_drag().

pub fn include_x(self, x: impl Into<f64>) -> Self

Expand bounds to include the given x value. For instance, to always show the y axis, call plot.include_x(0.0).

pub fn include_y(self, y: impl Into<f64>) -> Self

Expand bounds to include the given y value. For instance, to always show the x axis, call plot.include_y(0.0).

pub fn auto_bounds(self, auto_bounds: impl Into<Vec2b>) -> Self

Set whether the bounds should be automatically set based on data by default.

This is enabled by default.

pub fn auto_bounds_x(self) -> Self

👎Deprecated: Use auto_bounds instead

Expand bounds to fit all items across the x axis, including values given by include_x.

pub fn auto_bounds_y(self) -> Self

👎Deprecated: Use auto_bounds instead

Expand bounds to fit all items across the y axis, including values given by include_y.

pub fn legend(self, legend: Legend) -> Self

Show a legend including all named items.

pub fn show_background(self, show: bool) -> Self

Whether or not to show the background Rect.

Can be useful to disable if the plot is overlaid over existing content. Default: true.

pub fn show_axes(self, show: impl Into<Vec2b>) -> Self

Show axis labels and grid tick values on the side of the plot.

Default: true.

pub fn show_grid(self, show: impl Into<Vec2b>) -> Self

Show a grid overlay on the plot.

Default: true.

pub fn link_axis(self, group_id: impl Into<Id>, link: impl Into<Vec2b>) -> Self

Add this plot to an axis link group so that this plot will share the bounds with other plots in the same group. A plot cannot belong to more than one axis group.

pub fn link_cursor( self, group_id: impl Into<Id>, link: impl Into<Vec2b>, ) -> Self

Add this plot to a cursor link group so that this plot will share the cursor position with other plots in the same group. A plot cannot belong to more than one cursor group.

pub fn sharp_grid_lines(self, _enabled: bool) -> Self

👎Deprecated: This no longer has any effect and is always enabled.

Round grid positions to full pixels to avoid aliasing. Improves plot appearance but might have an undesired effect when shifting the plot bounds. Enabled by default.

pub fn reset(self) -> Self

Resets the plot.

pub fn x_axis_label(self, label: impl Into<WidgetText>) -> Self

Set the x axis label of the main X-axis.

Default: no label.

pub fn y_axis_label(self, label: impl Into<WidgetText>) -> Self

Set the y axis label of the main Y-axis.

Default: no label.

pub fn x_axis_position(self, placement: VPlacement) -> Self

Set the position of the main X-axis.

pub fn y_axis_position(self, placement: HPlacement) -> Self

Set the position of the main Y-axis.

pub fn x_axis_formatter( self, fmt: impl Fn(GridMark, &RangeInclusive<f64>) -> String + 'a, ) -> Self

Specify custom formatter for ticks on the main X-axis.

Arguments of fmt:

the grid mark to format

currently shown range on this axis.

pub fn y_axis_formatter( self, fmt: impl Fn(GridMark, &RangeInclusive<f64>) -> String + 'a, ) -> Self

Specify custom formatter for ticks on the main Y-axis.

Arguments of fmt:

the grid mark to format

currently shown range on this axis.

pub fn y_axis_min_width(self, min_width: f32) -> Self

Set the minimum width of the main y-axis, in ui points.

The width will automatically expand if any tickmark text is wider than this.

pub fn y_axis_width(self, digits: usize) -> Self

👎Deprecated: Use y_axis_min_width instead

Set the main Y-axis-width by number of digits

pub fn custom_x_axes(self, hints: Vec<AxisHints<'a>>) -> Self

Set custom configuration for X-axis

More than one axis may be specified. The first specified axis is considered the main axis.

pub fn custom_y_axes(self, hints: Vec<AxisHints<'a>>) -> Self

Set custom configuration for left Y-axis

More than one axis may be specified. The first specified axis is considered the main axis.

pub fn cursor_color(self, color: Color32) -> Self

Set custom cursor color.

You may set the color to Color32::TRANSPARENT to hide the cursors.

pub fn show<'b, R>( self, ui: &mut Ui, build_fn: impl FnOnce(&mut PlotUi<'b>) -> R + 'a, ) -> PlotResponse<R>

Interact with and add items to the plot and finally draw it.