egui_plot

Struct PlotImage

pub struct PlotImage { /* private fields */ }

An image in the plot.

Implementations

impl PlotImage

pub fn new( name: impl Into<String>, texture_id: impl Into<TextureId>, center_position: PlotPoint, size: impl Into<Vec2>, ) -> Self

Create a new image with position and size in plot coordinates.

pub fn uv(self, uv: impl Into<Rect>) -> Self

Select UV range. Default is (0,0) in top-left, (1,1) bottom right.

pub fn bg_fill(self, bg_fill: impl Into<Color32>) -> Self

A solid color to put behind the image. Useful for transparent images.

pub fn tint(self, tint: impl Into<Color32>) -> Self

Multiply image color with this. Default is WHITE (no tint).

pub fn rotate(self, angle: f64) -> Self

Rotate the image counter-clockwise around its center by an angle in radians.

pub fn name(self, name: impl ToString) -> Self

Name of this plot item.

This name will show up in the plot legend, if legends are turned on.

pub fn highlight(self, highlight: bool) -> Self

Highlight this plot item, typically by scaling it up.

If false, the item may still be highlighted via user interaction.

pub fn allow_hover(self, hovering: bool) -> Self

Allowed hovering this item in the plot. Default: true.

pub fn id(self, id: impl Into<Id>) -> Self

Sets the id of this plot item.

By default the id is determined from the name, but it can be explicitly set to a different value.