egui_plot

Struct VLine

pub struct VLine { /* private fields */ }

A vertical line in a plot, filling the full width

Implementations

impl VLine

pub fn new(name: impl Into<String>, x: impl Into<f64>) -> Self

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Add a stroke.

pub fn width(self, width: impl Into<f32>) -> Self

Stroke width. A high value means the plot thickens.

pub fn color(self, color: impl Into<Color32>) -> Self

Stroke color. Default is Color32::TRANSPARENT which means a color will be auto-assigned.

pub fn style(self, style: LineStyle) -> Self

Set the line’s style. Default is LineStyle::Solid.

pub fn name(self, name: impl ToString) -> Self

Name of this plot item.

This name will show up in the plot legend, if legends are turned on.

pub fn highlight(self, highlight: bool) -> Self

Highlight this plot item, typically by scaling it up.

If false, the item may still be highlighted via user interaction.

pub fn allow_hover(self, hovering: bool) -> Self

Allowed hovering this item in the plot. Default: true.

pub fn id(self, id: impl Into<Id>) -> Self

Sets the id of this plot item.

By default the id is determined from the name, but it can be explicitly set to a different value.