egui_plot

Function uniform_grid_spacer

pub fn uniform_grid_spacer<'a>(
    spacer: impl Fn(GridInput) -> [f64; 3] + 'a,
) -> Box<dyn Fn(GridInput) -> Vec<GridMark> + 'a>

Splits the grid into uniform-sized spacings (e.g. 100, 25, 1).

This function should return 3 positive step sizes, designating where the lines in the grid are drawn. Lines are thicker for larger step sizes. Ordering of returned value is irrelevant.

Why only 3 step sizes? Three is the number of different line thicknesses that egui typically uses in the grid. Ideally, those 3 are not hardcoded values, but depend on the visible range (accessible through GridInput).