egui_plot

Struct Line

pub struct Line<'a> { /* private fields */ }

A series of values forming a path.

Implementations

impl<'a> Line<'a>

pub fn new(series: impl Into<PlotPoints<'a>>) -> Self

pub fn highlight(self, highlight: bool) -> Self

Highlight this line in the plot by scaling up the line.

pub fn allow_hover(self, hovering: bool) -> Self

Allowed hovering this item in the plot. Default: true.

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Add a stroke.

pub fn width(self, width: impl Into<f32>) -> Self

Stroke width. A high value means the plot thickens.

pub fn color(self, color: impl Into<Color32>) -> Self

Stroke color. Default is Color32::TRANSPARENT which means a color will be auto-assigned.

pub fn fill(self, y_reference: impl Into<f32>) -> Self

Fill the area between this line and a given horizontal reference line.

pub fn fill_alpha(self, alpha: impl Into<f32>) -> Self

Set the fill area’s alpha channel. Default is 0.05.

pub fn style(self, style: LineStyle) -> Self

Set the line’s style. Default is LineStyle::Solid.

pub fn name(self, name: impl ToString) -> Self

Name of this line.

This name will show up in the plot legend, if legends are turned on.

Multiple plot items may share the same name, in which case they will also share an entry in the legend.

pub fn id(self, id: Id) -> Self

Set the line’s id which is used to identify it in the plot’s response.