egui_plot

Struct PlotUi

pub struct PlotUi<'a> { /* private fields */ }

Provides methods to interact with a plot while building it. It is the single argument of the closure provided to Plot::show. See Plot for an example of how to use it.

Implementations

impl<'a> PlotUi<'a>

pub fn ctx(&self) -> &Context

pub fn plot_bounds(&self) -> PlotBounds

The plot bounds as they were in the last frame. If called on the first frame and the bounds were not further specified in the plot builder, this will return bounds centered on the origin. The bounds do not change until the plot is drawn.

pub fn set_plot_bounds(&mut self, plot_bounds: PlotBounds)

Set the plot bounds. Can be useful for implementing alternative plot navigation methods.

pub fn translate_bounds(&mut self, delta_pos: Vec2)

Move the plot bounds. Can be useful for implementing alternative plot navigation methods.

pub fn auto_bounds(&self) -> Vec2b

Whether the plot axes were in auto-bounds mode in the last frame. If called on the first frame, this is the Plot’s default auto-bounds mode.

pub fn set_auto_bounds(&mut self, auto_bounds: impl Into<Vec2b>)

Set the auto-bounds mode for the plot axes.

pub fn response(&self) -> &Response

Can be used to check if the plot was hovered or clicked.

pub fn zoom_bounds(&mut self, zoom_factor: Vec2, center: PlotPoint)

Scale the plot bounds around a position in plot coordinates.

Can be useful for implementing alternative plot navigation methods.

The plot bounds are divided by zoom_factor, therefore:

zoom_factor < 1.0 zooms out, i.e., increases the visible range to show more data.

zoom_factor > 1.0 zooms in, i.e., reduces the visible range to show more detail.

pub fn zoom_bounds_around_hovered(&mut self, zoom_factor: Vec2)

Scale the plot bounds around the hovered position, if any.

Can be useful for implementing alternative plot navigation methods.

The plot bounds are divided by zoom_factor, therefore:

zoom_factor < 1.0 zooms out, i.e., increases the visible range to show more data.

zoom_factor > 1.0 zooms in, i.e., reduces the visible range to show more detail.

pub fn pointer_coordinate(&self) -> Option<PlotPoint>

The pointer position in plot coordinates. Independent of whether the pointer is in the plot area.

pub fn pointer_coordinate_drag_delta(&self) -> Vec2

The pointer drag delta in plot coordinates.

pub fn transform(&self) -> &PlotTransform

Read the transform between plot coordinates and screen coordinates.

pub fn screen_from_plot(&self, position: PlotPoint) -> Pos2

Transform the plot coordinates to screen coordinates.

pub fn plot_from_screen(&self, position: Pos2) -> PlotPoint

Transform the screen coordinates to plot coordinates.

pub fn add(&mut self, item: impl PlotItem + 'a)

Add an arbitrary item.

pub fn add_item(&mut self, item: Box<dyn PlotItem>)

Add an arbitrary item.

pub fn line(&mut self, line: Line<'a>)

Add a data line.

pub fn polygon(&mut self, polygon: Polygon<'a>)

Add a polygon. The polygon has to be convex.

pub fn text(&mut self, text: Text)

Add a text.

pub fn points(&mut self, points: Points<'a>)

Add data points.

pub fn arrows(&mut self, arrows: Arrows<'a>)

Add arrows.

pub fn image(&mut self, image: PlotImage)

Add an image.

pub fn hline(&mut self, hline: HLine)

Add a horizontal line. Can be useful e.g. to show min/max bounds or similar. Always fills the full width of the plot.

pub fn vline(&mut self, vline: VLine)

Add a vertical line. Can be useful e.g. to show min/max bounds or similar. Always fills the full height of the plot.

pub fn box_plot(&mut self, box_plot: BoxPlot)

Add a box plot diagram.

pub fn bar_chart(&mut self, chart: BarChart)

Add a bar chart.