egui_plot

Struct VLine

pub struct VLine { /* private fields */ }

A vertical line in a plot, filling the full width

Implementations

impl VLine

pub fn new(x: impl Into<f64>) -> Self

pub fn highlight(self, highlight: bool) -> Self

Highlight this line in the plot by scaling up the line.

pub fn allow_hover(self, hovering: bool) -> Self

Allowed hovering this item in the plot. Default: true.

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Add a stroke.

pub fn width(self, width: impl Into<f32>) -> Self

Stroke width. A high value means the plot thickens.

pub fn color(self, color: impl Into<Color32>) -> Self

Stroke color. Default is Color32::TRANSPARENT which means a color will be auto-assigned.

pub fn style(self, style: LineStyle) -> Self

Set the lineโs style. Default is LineStyle::Solid.

pub fn name(self, name: impl ToString) -> Self

Name of this vertical line.

This name will show up in the plot legend, if legends are turned on.

Multiple plot items may share the same name, in which case they will also share an entry in the legend.

pub fn id(self, id: Id) -> Self

Set the lineโs id which is used to identify it in the plotโs response.